+++ src/DataFilters/Grammar/Syntax/ConstantValueExpression.cs
@@ -19,11 +19,10 @@ namespace DataFilters.Grammar.Syntax
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is <see cref="string.Empty"/> or <paramref name="value"/> is not currently supported
         /// </exception>
         /// <exception cref="ArgumentNullException"><paramref name="value"/> is <c>null</c>.</exception>
-        public ConstantValueExpression(object value)
+        private ConstantValueExpression(object value)
         {
             Value = value switch
             {
-                null => throw new ArgumentNullException(nameof(value)),
                 short shortValue => shortValue,
                 string stringValue when stringValue.Length == 0 => throw new ArgumentOutOfRangeException(nameof(value), value, $"{nameof(value)} cannot be empty"),
                 string stringValue => stringValue,
@@ -34,9 +33,59 @@ namespace DataFilters.Grammar.Syntax
                 bool boolValue => boolValue,
                 byte byteValue => byteValue,
                 int intValue => intValue,
+                char chrValue => chrValue,
+                null => throw new ArgumentNullException(nameof(value)),
                 _ => throw new ArgumentOutOfRangeException(nameof(value), $"Unsupported type '{value.GetType()}' for {nameof(ConstantValueExpression)}")
             };
         }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="string"/> <paramref name="value"/>.
+        /// </summary>
+        /// <param name="value"></param>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is <see cref="string.Empty"/> or <paramref name="value"/> is not currently supported
+        /// </exception>
+        /// <exception cref="ArgumentNullException"><paramref name="value"/> is <c>null</c>.</exception>
+        public ConstantValueExpression(string value) : this((object)value) { }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="DateTime"/> <paramref name="value"/>.
+        /// </summary>
+        /// <param name="value"></param>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is <see cref="string.Empty"/> or <paramref name="value"/> is not currently supported
+        /// </exception>
+        /// <exception cref="ArgumentNullException"><paramref name="value"/> is <c>null</c>.</exception>
+        public ConstantValueExpression(DateTime value) : this((object)value) { }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="DateTimeOffset"/> <paramref name="value"/>.
+        /// </summary>
+        public ConstantValueExpression(DateTimeOffset value) : this((object)value) { }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="long"/> <paramref name="value"/>.
+        /// </summary>
+        public ConstantValueExpression(long value) : this((object)value) { }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="int"/> <paramref name="value"/>.
+        /// </summary>
+        public ConstantValueExpression(int value) : this((object)value) { }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="byte"/> <paramref name="value"/>.
+        /// </summary>
+        public ConstantValueExpression(byte value) : this((object)value) { }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="Guid"/> <paramref name="value"/>.
+        /// </summary>
+        public ConstantValueExpression(Guid value) : this((object)value) { }
+
+        /// <summary>
+        /// Builds a new <see cref="ConstantValueExpression"/> that holds the specified <see cref="Guid"/> <paramref name="value"/>.
+        /// </summary>
+        public ConstantValueExpression(bool value) : this((object)value) { }
 
         ///<inheritdoc/>
         public bool Equals(ConstantValueExpression other) => Equals(Value, other?.Value);