+++ test/DataFilters.UnitTests/Helpers/ExpressionsGenerators.cs
@@ -48,21 +48,21 @@ namespace DataFilters.UnitTests.Helpers
 
         public static Arbitrary<ConstantValueExpression> GenerateConstantValueExpression()
         {
-            IList<Gen<object>> generators = new List<Gen<object>>
+            IList<Gen<ConstantValueExpression>> generators = new List<Gen<ConstantValueExpression>>
             {
-                Arb.Default.Bool().Generator.Select(item => (object) item),
-                Arb.Default.String().Generator.Where(item => !string.IsNullOrEmpty(item)).Select(item => (object) item),
-                Arb.Default.Int16().Generator.Select(item => (object) item),
-                Arb.Default.Int32().Generator.Select(item => (object) item),
-                Arb.Default.Int64().Generator.Select(item => (object) item),
-                Arb.Default.DateTime().Generator.Select(item => (object) item),
-                Arb.Default.DateTimeOffset().Generator.Select(item => (object) item),
-                Arb.Default.Byte().Generator.Select(item => (object) item),
-                Arb.Default.Guid().Generator.Select(item => (object) item)
+                Arb.Default.Bool().Generator.Select(value => new ConstantValueExpression(value)),
+                Arb.Default.NonWhiteSpaceString().Generator.Select(value => new ConstantValueExpression(value.Item)),
+                Arb.Default.Int16().Generator.Select(value => new ConstantValueExpression(value)),
+                Arb.Default.Int32().Generator.Select(value => new ConstantValueExpression(value)),
+                Arb.Default.Int64().Generator.Select(value => new ConstantValueExpression(value)),
+                Arb.Default.DateTime().Generator.Select(value => new ConstantValueExpression(value)),
+                Arb.Default.DateTimeOffset().Generator.Select(value => new ConstantValueExpression(value)),
+                Arb.Default.Byte().Generator.Select(value => new ConstantValueExpression(value)),
+                Arb.Default.Guid().Generator.Select(value => new ConstantValueExpression(value))
             };
 
-            return Gen.OneOf(generators).Select(generatedValue => new ConstantValueExpression(generatedValue))
-                       .ToArbitrary();
+            return Gen.OneOf(generators)
+                      .ToArbitrary();
         }
 
         public static Arbitrary<DurationExpression> GenerateDurationExpression()
@@ -120,8 +120,8 @@ namespace DataFilters.UnitTests.Helpers
             return Gen.OneOf(generators)
                       .Zip(Arb.Default.Bool().Generator)
                       .Two()
-                      .Select(tuple => new RangeExpression(min: new BoundaryExpression(expression : tuple.Item1.Item1, included: tuple.Item1.Item2),
-                                                           max: new BoundaryExpression(expression : tuple.Item2.Item1, included: tuple.Item2.Item2)))
+                      .Select(tuple => new RangeExpression(min: new BoundaryExpression(expression: tuple.Item1.Item1, included: tuple.Item1.Item2),
+                                                           max: new BoundaryExpression(expression: tuple.Item2.Item1, included: tuple.Item2.Item2)))
                       .ToArbitrary();
         }
 
@@ -130,20 +130,40 @@ namespace DataFilters.UnitTests.Helpers
                           .Convert(convertTo: input => new ContainsExpression(input.Item),
                                    convertFrom: expression => NonEmptyString.NewNonEmptyString(expression.Value));
 
-        public static Arbitrary<RegexValue> GenerateRegexValues()
+        public static Arbitrary<BracketValue> GenerateRegexValues()
         {
-            Arbitrary<char> letterOrNumberCharacter = Arb.Default.Char()
-                                                 .Filter(chr => (('a' <= chr && chr <= 'Z') || ('0' <= chr && chr <= '9')));
-            Gen<RegexValue> regexRangeGenerator = letterOrNumberCharacter
-                                     .Generator
-                                     .Two()
-                                     .Where(tuple => tuple.Item1 <= tuple.Item2)
-                                     .Select(tuple => (RegexValue) new RegexRangeValue(tuple.Item1, tuple.Item2));
-            Gen<RegexValue> regexConstantGenerator = letterOrNumberCharacter
-                .Generator
-                .Select(chr => (RegexValue) new RegexConstantValue(chr));
+            Gen<BracketValue> regexRangeGenerator = BuildRegexRangeValueGenerator().Convert(range => (BracketValue)range,
+                                                                                            bracket => (RangeBracketValue)bracket)
+                                                                                                          .Generator;
+            Gen<BracketValue> regexConstantGenerator = BuildRegexConstantValueGenerator().Convert(range => (BracketValue)range,
+                                                                                                  bracket => (ConstantBracketValue)bracket)
+                                                                                         .Generator;
 
             return Gen.OneOf(regexConstantGenerator, regexRangeGenerator).ToArbitrary();
+        }
+
+        public static Arbitrary<ConstantBracketValue> BuildRegexConstantValueGenerator() => Arb.Default.String()
+                                                                                                   .Filter(input => !string.IsNullOrWhiteSpace(input)
+                                                                                                                    && input.Length > 1
+                                                                                                                    && input.All(chr => char.IsLetterOrDigit(chr)))
+                                                                                                   .Generator
+                                                                                                   .Select(item => new ConstantBracketValue(item))
+                                                                                                   .ToArbitrary();
+
+        public static Arbitrary<RangeBracketValue> BuildRegexRangeValueGenerator()
+        {
+            return Arb.Default.Char()
+                              .Filter(chr => char.IsLetterOrDigit(chr)).Generator
+                              .Two()
+                              .Select(tuple => (start: tuple.Item1, end: tuple.Item2))
+                              .Where(tuple => (tuple.start < tuple.end) && (TupleContainsLetter(tuple) || TupleContainsDigits(tuple)))
+                              .Select(tuple => new RangeBracketValue(tuple.start, tuple.end))
+                              .ToArbitrary();
+
+            static bool TupleContainsLetter((char start, char end) tuple) => char.IsLetter(tuple.start) && char.IsLetter(tuple.end)
+                                                                            && (( char.IsLower(tuple.start) && char.IsLower(tuple.end)  )|| (char.IsUpper(tuple.start) && char.IsUpper(tuple.end)));
+
+            static bool TupleContainsDigits((char start, char end) tuple) => char.IsDigit(tuple.start) && char.IsDigit(tuple.end);
         }
     }
 }